Disentangle Mac impl from Linux/BSD impl
This should provide a working Mac implementation. There is some overlap
between the Mac and Linux/BSD implementations, and it is possible that a
helper function could be factored out from both to allow both implementations
to share code. However, I think the Mac version is easier to understand if
it's completely separate.