Fix problem with denormalized paths in executable_path_test.c
On Mac, I have observed that cwk_path_get_executable_path
returns a denormalized path in some situations; this causes
the tests to fail.

One solution to this issue is to always normalize the path
that this function returns; this would require an extra buffer
and additional overhead. The user may not care if the path is
denormalized or not; in this case, that overhead would be
unnecessary.

Instead, I changed the documentation to say that the path may not
be normalized, and changed the test code to normalize the path
before comparison.