Move executable_path tests into ctest
This moves the cwalktest_exe_path binary out of the global scope and
into the 'ENABLE_TESTS' scope, so that the binary is only built when
tests are enabled. It also uses 'add_test' so that CMake/CTest/Make
know about cwalktest_exe_path, and know how to run it using
'test_executable_paths.sh'. This allows the test to run when you use
'make test' or 'ctest'.